requirements.txt
# Core dependencies
redis>=5.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Async and networking
aiohttp>=3.9.0
websockets>=12.0
python-socketio[asyncio_client]>=5.7.0

# Logging
structlog>=23.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
